refactor(app): move probe tests beside readiness.go and pin per_tenant
Review fix round 1.

- prealloc: names is built with make([]string, 0, len(probes)).
- readyCheck's db_stats comment no longer claims the other components
  report an empty object: every classic kind now describes itself, so the
  dbStatus.Status == "" branch only covers a hand-built healthProbes slice
  with no database description. Behavior unchanged.
- per_tenant is now a rule every leased kind shares, so messagingProbe and
  cacheProbe get the relabel tests the database kind already had; both fail
  (not_configured) when the description's perTenant flag is forced false.
- createTestMessagingManagerWithStats had one caller passing nil, leaving
  its pre-population block dead: it becomes
  createTestMessagingManagerWithReadyClient(t).
- Source-to-test 1:1: the tests exercising readiness.go symbols move from
  health_test.go to readiness_test.go, taking the fixtures only they use
  (newDBManagerFor, createWarmCacheManagerWithHungPing, the messaging
  fixtures). health_test.go keeps TestGetStatsOrEmpty plus the manager
  fixtures and /ready body constants shared with app, lifecycle and
  debug-health tests.
